fix(folders): change files location
BREAKING CHANGE: organize files using same principle as OpenGeode v2
