+++ include/geode/geosciences/representation/builder/structural_model_builder.h
@@ -26,12 +26,12 @@
 #include <geode/model/mixin/builder/add_components_builders.h>
 #include <geode/model/representation/builder/brep_builder.h>
 
-#include <geode/geosciences/builder/faults_builder.h>
-#include <geode/geosciences/builder/horizons_builder.h>
-#include <geode/geosciences/core/fault.h>
-#include <geode/geosciences/core/faults.h>
-#include <geode/geosciences/core/horizon.h>
-#include <geode/geosciences/core/horizons.h>
+#include <geode/geosciences/mixin/builder/faults_builder.h>
+#include <geode/geosciences/mixin/builder/horizons_builder.h>
+#include <geode/geosciences/mixin/core/fault.h>
+#include <geode/geosciences/mixin/core/faults.h>
+#include <geode/geosciences/mixin/core/horizon.h>
+#include <geode/geosciences/mixin/core/horizons.h>
 
 namespace geode
 {
