+++ include/geode/geosciences/representation/io/structural_model_input.h
