+++ src/geode/geosciences/CMakeLists.txt
@@ -1,34 +1,34 @@
 
 set(sources
-    "${lib_source_dir}/builder/structural_model_builder.cpp"
-    "${lib_source_dir}/builder/faults_builder.cpp"
-    "${lib_source_dir}/builder/horizons_builder.cpp"
-    "${lib_source_dir}/core/bitsery_archive.cpp"
-    "${lib_source_dir}/core/fault.cpp"
-    "${lib_source_dir}/core/faults.cpp"
-    "${lib_source_dir}/core/horizon.cpp"
-    "${lib_source_dir}/core/horizons.cpp"
-    "${lib_source_dir}/io/structural_model_input.cpp"
-    "${lib_source_dir}/io/structural_model_output.cpp"
+    "${lib_source_dir}/mixin/builder/faults_builder.cpp"
+    "${lib_source_dir}/mixin/builder/horizons_builder.cpp"
+    "${lib_source_dir}/mixin/core/bitsery_archive.cpp"
+    "${lib_source_dir}/mixin/core/fault.cpp"
+    "${lib_source_dir}/mixin/core/faults.cpp"
+    "${lib_source_dir}/mixin/core/horizon.cpp"
+    "${lib_source_dir}/mixin/core/horizons.cpp"
+    "${lib_source_dir}/representation/builder/structural_model_builder.cpp"
+    "${lib_source_dir}/representation/io/structural_model_input.cpp"
+    "${lib_source_dir}/representation/io/structural_model_output.cpp"
 )
 set(public_headers
     "${lib_include_dir}/common.h"
-    "${lib_include_dir}/builder/structural_model_builder.h"
-    "${lib_include_dir}/builder/faults_builder.h"
-    "${lib_include_dir}/builder/horizons_builder.h"
-    "${lib_include_dir}/core/bitsery_archive.h"
-    "${lib_include_dir}/core/fault.h"
-    "${lib_include_dir}/core/faults.h"
-    "${lib_include_dir}/core/horizon.h"
-    "${lib_include_dir}/core/horizons.h"
-    "${lib_include_dir}/core/structural_model.h"
-    "${lib_include_dir}/io/structural_model_input.h"
-    "${lib_include_dir}/io/structural_model_output.h"
+    "${lib_include_dir}/mixin/builder/faults_builder.h"
+    "${lib_include_dir}/mixin/builder/horizons_builder.h"
+    "${lib_include_dir}/mixin/core/bitsery_archive.h"
+    "${lib_include_dir}/mixin/core/fault.h"
+    "${lib_include_dir}/mixin/core/faults.h"
+    "${lib_include_dir}/mixin/core/horizon.h"
+    "${lib_include_dir}/mixin/core/horizons.h"
+    "${lib_include_dir}/representation/builder/structural_model_builder.h"
+    "${lib_include_dir}/representation/core/structural_model.h"
+    "${lib_include_dir}/representation/io/structural_model_input.h"
+    "${lib_include_dir}/representation/io/structural_model_output.h"
 )
 set(advanced_headers
-    "${lib_include_dir}/core/detail/components_storage.h"
-    "${lib_include_dir}/io/detail/geode_structural_model_input.h"
-    "${lib_include_dir}/io/detail/geode_structural_model_output.h"
+    "${lib_include_dir}/mixin/core/detail/components_storage.h"
+    "${lib_include_dir}/representation/io/detail/geode_structural_model_input.h"
+    "${lib_include_dir}/representation/io/detail/geode_structural_model_output.h"
 )
 target_link_libraries(${target_name}
     PRIVATE
