+++ upgrade-guide.md
@@ -1,0 +1,12 @@
+# Upgrade Guide
+
+## Upgrading from OpenGeode-Geosciences v1.x.x to v2.0.0
+
+### Motivations
+
+Following the same organization principle then OpenGeode v2.
+
+
+### Breaking Changes
+
+- Files have been reorganized by adding two folders: mixin and representation.
